Move tasks into their own package
