+++ cmd/mailroom/main.go
@@ -13,15 +13,15 @@ import (
 	"github.com/nyaruka/mailroom/config"
 	"github.com/sirupsen/logrus"
 
-	_ "github.com/nyaruka/mailroom/broadcasts"
-	_ "github.com/nyaruka/mailroom/campaigns"
-	_ "github.com/nyaruka/mailroom/expirations"
 	_ "github.com/nyaruka/mailroom/hooks"
-	_ "github.com/nyaruka/mailroom/interrupts"
 	_ "github.com/nyaruka/mailroom/ivr"
-	_ "github.com/nyaruka/mailroom/starts"
-	_ "github.com/nyaruka/mailroom/stats"
-	_ "github.com/nyaruka/mailroom/timeouts"
+	_ "github.com/nyaruka/mailroom/tasks/broadcasts"
+	_ "github.com/nyaruka/mailroom/tasks/campaigns"
+	_ "github.com/nyaruka/mailroom/tasks/expirations"
+	_ "github.com/nyaruka/mailroom/tasks/interrupts"
+	_ "github.com/nyaruka/mailroom/tasks/starts"
+	_ "github.com/nyaruka/mailroom/tasks/stats"
+	_ "github.com/nyaruka/mailroom/tasks/timeouts"
 
 	_ "github.com/nyaruka/mailroom/web/docs"
 	_ "github.com/nyaruka/mailroom/web/expression"
